Feature/backend update user details

+++ backend/src/main/java/com/mynt/banking/user/token/TokenRepository.java
@@ -7,14 +7,14 @@ import java.util.List;
 import java.util.Optional;
 
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
 
     @Query(value = """
       select t from Token t inner join User u\s
       on t.user.id = u.id\s
       where u.id = :id and (t.expired = false or t.revoked = false)\s
       """)
-    List<Token> findAllValidTokenByUser(Integer id);
+    List<Token> findAllValidTokenByUser(Long id);
 
     Optional<Token> findByToken(String token);
 }
